+++ ProjectPortfolio/Services/SystemUserService.cs
@@ -1,11 +1,12 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
 using ProjectPortfolio.Data;
+using ProjectPortfolio.Enumerators;
 using ProjectPortfolio.Models;
 
 namespace ProjectPortfolio.Services
 {
-    internal class SystemUserService(ISystemUserRepository repository) : ISystemUserService
+    internal class SystemUserService(ISystemUserRepository repository, IIssueRepository issueRepository) : ISystemUserService
     {
         public async Task<SystemUserModel> CreateAsync(SystemUserModel model)
         {
@@ -17,14 +18,17 @@ namespace ProjectPortfolio.Services
             return result;
         }
 
-        public Task<SystemUserModel> UpdateAsync(SystemUserModel model)
+        public async Task<SystemUserModel> UpdateAsync(SystemUserModel model)
         {
-            throw new NotImplementedException();
+            var result = await repository.UpdateAsync(model);   
+            return result;
         }
 
         public async Task DeleteAsync(Guid id)
         {
-            var query = repository.GetAll().Where(e => e.Id == id).FirstOrDefaultAsync();
+            var issues = await issueRepository.GetAll().AsNoTracking().Where(e => e.AttendantId == id && e.Status != IssueStatusEnum.Closed).ToListAsync();
+            if (issues.Count > 0)
+                throw new Exception("Usuário está vinculado a atividade ativa e não pode ser deletado.");
 
             await repository.DeleteAsync(id);
         }
@@ -32,7 +36,7 @@ namespace ProjectPortfolio.Services
         private static void SystemUserValidator(SystemUserModel model)
         {
             if (model.Name.Length < 3 || model.Name.Length > 35 || model.Name.IsNullOrEmpty())
-                throw new Exception("Nome deve ter entre {0} e {1} caracteres.");
+                throw new Exception("Nome deve ter entre 3 e 35 caracteres.");
             if (model.Surname.Length < 3 || model.Surname.Length > 100 || model.Surname.IsNullOrEmpty())
                 throw new Exception("O sobrenome deve ter entre 3 e 100 caracteres.");
             if(model.UserName.Length < 3 || model.UserName.Length > 50 || model.Name.IsNullOrEmpty())
